initial commit; added interior_point.m
interior_point.m will contain a MATLAB function
that will find the minimum of a quadratic function F
given a set of constraints and a starting position (x, y),
where x is the vector that will optimize F and y is the
lagrange multiplier vector.

currently, the code is a simple script.
the LP portion of the script has been tested.
the QP portion has not been tested.
once testing is complete, this script will be
adapted into a function.